--- frontend/src/components/FeedArea/FeedArea.js
@@ -1,13 +1,0 @@
-import ImagesSection from 'components/UploadedImagesSection/UploadedImagesSection';
-import { FeedArea, FeedTitleArea } from 'components/FeedArea/FeedArea.styles';
-
-const Feed = () => (
-  <FeedArea>
-    <FeedTitleArea>
-      <h2>Feed</h2>
-    </FeedTitleArea>
-    <ImagesSection />
-  </FeedArea>
-);
-
-export default Feed;
