--- frontend/src/components/ImageDetails/ImageDetails.js
@@ -1,16 +1,0 @@
-import {Details, Ul, Li} from "./ImageDetails.styles";
-
-const ImageDetails = () => (
-  <Details>
-    <Ul>
-      Objects
-      <Li>obj1</Li>
-    </Ul>
-    <Ul>
-      Labels
-      <Li>lbl1</Li>
-    </Ul>
-  </Details>
-);
-
-export default ImageDetails;
