--- frontend/src/components/UploadImageArea/UploadImage.js
@@ -1,10 +1,0 @@
-import { StyledUploadImage } from './UploadImage.styles';
-
-const UploadImage = () => (
-  <StyledUploadImage>
-    <p>Upload an image or drag and drop</p>
-    <div></div>
-  </StyledUploadImage>
-);
-
-export default UploadImage;
