+++ frontend/src/views/HomePage/HomePage.js
@@ -1,0 +1,20 @@
+import UploadImage from 'views/HomePage/components/UploadImageArea/UploadImage';
+import Feed from 'views/HomePage/components/FeedArea/FeedArea';
+import Header from 'components/Header/Header';
+import Footer from 'components/Footer/Footer';
+import { Wrapper } from './HomePage.styles'
+
+const Home = () => {
+  return (
+    <>
+      <Wrapper>
+        <Header />
+        <UploadImage/>
+        <Feed />
+      </Wrapper>
+      <Footer />
+    </>
+  );
+};
+
+export default Home;
