+++ frontend/src/views/HomePage/HomePage.styles.js
@@ -8,5 +8,5 @@ export const Wrapper = styled.div`
   max-width: 100%;
   margin: 0 auto;
   min-height: 100vh;
-  background-color: ${({ theme }) => theme.colors.primary};
+  /* background-color: ${({ theme }) => theme.colors.primary}; */
 `;
