+++ frontend/src/views/HomePage/components/FeedArea/FeedArea.js
@@ -1,0 +1,13 @@
+import ImagesSection from 'views/HomePage/components/UploadedImagesSection/UploadedImagesSection';
+import { FeedArea, FeedTitleArea } from 'views/HomePage/components/FeedArea/FeedArea.styles';
+
+const Feed = () => (
+  <FeedArea>
+    <FeedTitleArea>
+      <h2>Feed</h2>
+    </FeedTitleArea>
+    <ImagesSection/>
+  </FeedArea>
+);
+
+export default Feed;
