+++ frontend/src/views/HomePage/components/UploadImageArea/UploadImage.js
@@ -1,0 +1,14 @@
+import { StyledUploadImage } from './UploadImage.styles';
+
+const UploadImage = () => {
+
+  return (
+    <StyledUploadImage>
+      <p>Upload an image or drag and drop</p>
+      <div>
+      </div>
+    </StyledUploadImage>
+  );
+};
+
+export default UploadImage;
