+++ frontend/src/views/HomePage/components/UploadedImage/UploadedImage.styles.js
