+++ frontend/src/views/HomePage/components/UploadedImagesSection/UploadedImagesSection.styles.js
