+++ frontend/src/views/ImageDetailsPage/components/ImageDetails/ImageDetails.js
@@ -1,0 +1,18 @@
+import { Details, Ul, Li } from './ImageDetails.styles';
+
+const ImageDetails = () => {
+  return (
+    <Details>
+      <Ul>
+        Objects
+        <Li>Obj1</Li>
+      </Ul>
+      <Ul>
+        Labels
+        <Li>lbl1</Li>
+      </Ul>
+    </Details>
+  );
+};
+
+export default ImageDetails;
